+++ share/translations/keepassx_en.ts
@@ -1114,6 +1114,14 @@ chrome-laptop.</source>
         <source>Column %1</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>TOTP</source>
+        <translation type="unfinished">TOTP</translation>
+    </message>
+    <message>
+        <source>Icon</source>
+        <translation type="unfinished">Icon</translation>
+    </message>
 </context>
 <context>
     <name>CsvParserModel</name>
@@ -1217,20 +1225,6 @@ Backup database located at %2</source>
     <message>
         <source>Refresh</source>
         <translation>Refresh</translation>
-    </message>
-    <message>
-        <source>Legacy key file format</source>
-        <translation>Legacy key file format</translation>
-    </message>
-    <message>
-        <source>You are using a legacy key file format which may become
-unsupported in the future.
-
-Please consider generating a new key file.</source>
-        <translation>You are using a legacy key file format which may become
-unsupported in the future.
-
-Please consider generating a new key file.</translation>
     </message>
     <message>
         <source>Don&apos;t show this warning again</source>
@@ -1355,6 +1349,14 @@ If you do not have a key file, please leave the field empty.</source>
     </message>
     <message>
         <source>Select hardware key…</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Old key file format</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>You are using an old key file format which KeePassXC may&lt;br&gt;stop supporting in the future.&lt;br&gt;&lt;br&gt;Please consider generating a new key file by going to:&lt;br&gt;&lt;strong&gt;Database / Database Security / Change Key File.&lt;/strong&gt;&lt;br&gt;</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
@@ -2322,6 +2324,15 @@ Disable safe saves and try again?</translation>
     </message>
     <message>
         <source>[PROTECTED] Press Reveal to view or edit</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Invalid Entry</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>An external merge operation has invalidated this entry.
+Unfortunately, any changes made have been lost.</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
@@ -3465,11 +3476,6 @@ Are you sure to add this file?</source>
         <translation>[PROTECTED]</translation>
     </message>
     <message>
-        <source>&lt;b&gt;%1&lt;/b&gt;: %2</source>
-        <comment>attributes line</comment>
-        <translation>&lt;b&gt;%1&lt;/b&gt;: %2</translation>
-    </message>
-    <message>
         <source>Enabled</source>
         <translation>Enabled</translation>
     </message>
@@ -3488,6 +3494,15 @@ Are you sure to add this file?</source>
     <message>
         <source>Advanced</source>
         <translation type="unfinished">Advanced</translation>
+    </message>
+    <message>
+        <source>Default Sequence</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>&lt;tr&gt;&lt;td&gt;&lt;b&gt;%1&lt;/b&gt;:&lt;/td&gt;&lt;td&gt;%2&lt;/td&gt;&lt;/tr&gt;</source>
+        <comment>attributes line</comment>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
@@ -3523,22 +3538,7 @@ Are you sure to add this file?</source>
     </message>
 </context>
 <context>
-    <name>FdoSecrets::Item</name>
-    <message>
-        <source>Entry &quot;%1&quot; from database &quot;%2&quot; was used by %3</source>
-        <translation type="unfinished"></translation>
-    </message>
-</context>
-<context>
     <name>FdoSecrets::Service</name>
-    <message numerus="yes">
-        <source>%n Entry(s) was used by %1</source>
-        <comment>%1 is the name of an application</comment>
-        <translation type="unfinished">
-            <numerusform></numerusform>
-            <numerusform></numerusform>
-        </translation>
-    </message>
     <message>
         <source>Failed to register DBus service at %1.&lt;br/&gt;</source>
         <translation type="unfinished"></translation>
@@ -4363,10 +4363,6 @@ If this reoccurs, then your database file may be corrupt.</source>
         <translation>&lt;p&gt;You can add a key file containing random bytes for additional security.&lt;/p&gt;&lt;p&gt;You must keep it secret and never lose it or you will be locked out!&lt;/p&gt;</translation>
     </message>
     <message>
-        <source>Legacy key file format</source>
-        <translation>Legacy key file format</translation>
-    </message>
-    <message>
         <source>Error loading the key file &apos;%1&apos;
 Message: %2</source>
         <translation>Error loading the key file &apos;%1&apos;
@@ -4434,10 +4430,11 @@ Are you sure you want to continue with this file?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>You are using a legacy key file format which may become
-unsupported in the future.
-
-Generate a new key file in the database security settings.</source>
+        <source>Old key file format</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>You selected a key file in an old format which KeePassXC&lt;br&gt;may stop supporting in the future.&lt;br&gt;&lt;br&gt;Please consider generating a new key file instead.</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
@@ -5785,16 +5782,6 @@ Expect some bugs and minor issues, this version is not meant for production use.
         <translation>Perform advanced analysis on the password.</translation>
     </message>
     <message>
-        <source>WARNING: You are using a legacy key file format which may become
-unsupported in the future.
-
-Please consider generating a new key file.</source>
-        <translation>WARNING: You are using a legacy key file format which may become
-unsupported in the future.
-
-Please consider generating a new key file.</translation>
-    </message>
-    <message>
         <source>
 
 Available commands:
@@ -6183,10 +6170,6 @@ Available commands:
     <message>
         <source>%1: (row, col) %2,%3</source>
         <translation>%1: (row, col) %2,%3</translation>
-    </message>
-    <message>
-        <source>Argon2 (KDBX 4 – recommended)</source>
-        <translation>Argon2 (KDBX 4 – recommended)</translation>
     </message>
     <message>
         <source>AES-KDF (KDBX 4)</source>
@@ -6765,10 +6748,6 @@ Kernel: %3 %4</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Argon2 (%1 rounds, %2 KB)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <source>AES 256-bit</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6802,6 +6781,45 @@ Kernel: %3 %4</source>
     </message>
     <message>
         <source>path to a custom local config file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>WARNING: You are using an old key file format which KeePassXC may
+stop supporting in the future.
+
+Please consider generating a new key file.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Argon2%1 (%2 rounds, %3 KB)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Argon2d (KDBX 4 – recommended)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Argon2id (KDBX 4)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>TOTP</source>
+        <translation type="unfinished">TOTP</translation>
+    </message>
+    <message>
+        <source>Icon</source>
+        <translation type="unfinished">Icon</translation>
+    </message>
+    <message>
+        <source>Unsupported key file version: %1</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Checksum mismatch! Key file may be corrupt.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Unexpected key file data! Key file may be corrupt.</source>
         <translation type="unfinished"></translation>
     </message>
 </context>