+++ src/core/Entry.h
@@ -161,6 +161,8 @@ public:
         CloneNewUuid = 1, // generate a random uuid for the clone
         CloneResetTimeInfo = 2, // set all TimeInfo attributes to the current time
         CloneIncludeHistory = 4, // clone the history items
+        CloneDefault = CloneNewUuid | CloneResetTimeInfo,
+        CloneCopy = CloneNewUuid | CloneResetTimeInfo | CloneIncludeHistory,
         CloneRenameTitle = 8, // add "-Clone" after the original title
         CloneUserAsRef = 16, // Add the user as a reference to the original entry
         ClonePassAsRef = 32, // Add the password as a reference to the original entry
@@ -210,7 +212,6 @@ public:
     static const int ResolveMaximumDepth;
     static const QString AutoTypeSequenceUsername;
     static const QString AutoTypeSequencePassword;
-    static CloneFlags DefaultCloneFlags;
 
     /**
      * Creates a duplicate of this entry except that the returned entry isn't
@@ -218,7 +219,7 @@ public:
      * Note that you need to copy the custom icons manually when inserting the
      * new entry into another database.
      */
-    Entry* clone(CloneFlags flags = DefaultCloneFlags) const;
+    Entry* clone(CloneFlags flags = CloneDefault) const;
     void copyDataFrom(const Entry* other);
     QString maskPasswordPlaceholders(const QString& str) const;
     Entry* resolveReference(const QString& str) const;