+++ src/gui/EntryPreviewWidget.cpp
@@ -286,14 +286,18 @@ void EntryPreviewWidget::updateEntryAdvancedTab()
 
     setTabEnabled(m_ui->entryTabWidget, m_ui->entryAdvancedTab, hasAttributes || hasAttachments);
     if (hasAttributes) {
-        QString attributesText;
+        QString attributesText("<table>");
         for (const QString& key : customAttributes) {
-            QString value = m_currentEntry->attributes()->value(key);
+            QString value;
             if (m_currentEntry->attributes()->isProtected(key)) {
                 value = "<i>" + tr("[PROTECTED]") + "</i>";
+            } else {
+                value = m_currentEntry->attributes()->value(key).toHtmlEscaped();
+                value.replace('\n', QLatin1String("<br/>"));
             }
-            attributesText.append(tr("<b>%1</b>: %2", "attributes line").arg(key, value).append("<br/>"));
+            attributesText.append(tr("<tr><td><b>%1</b>:</td><td>%2</td></tr>", "attributes line").arg(key, value));
         }
+        attributesText.append("</table>");
         m_ui->entryAttributesEdit->setText(attributesText);
     }
 
@@ -303,6 +307,8 @@ void EntryPreviewWidget::updateEntryAdvancedTab()
 void EntryPreviewWidget::updateEntryAutotypeTab()
 {
     Q_ASSERT(m_currentEntry);
+
+    m_ui->entrySequenceLabel->setText(m_currentEntry->effectiveAutoTypeSequence());
     m_ui->entryAutotypeTree->clear();
     QList<QTreeWidgetItem*> items;
     const AutoTypeAssociations* autotypeAssociations = m_currentEntry->autoTypeAssociations();
@@ -314,7 +320,7 @@ void EntryPreviewWidget::updateEntryAutotypeTab()
     }
 
     m_ui->entryAutotypeTree->addTopLevelItems(items);
-    setTabEnabled(m_ui->entryTabWidget, m_ui->entryAutotypeTab, !items.isEmpty());
+    setTabEnabled(m_ui->entryTabWidget, m_ui->entryAutotypeTab, m_currentEntry->autoTypeEnabled());
 }
 
 void EntryPreviewWidget::updateGroupHeaderLine()