+++ src/gui/osutils/nixutils/NixUtils.cpp
@@ -127,10 +127,12 @@ void NixUtils::setLaunchAtStartup(bool enable)
                << QStringLiteral("StartupNotify=true") << '\n'
                << QStringLiteral("Terminal=false") << '\n'
                << QStringLiteral("Type=Application") << '\n'
-               << QStringLiteral("Version=1.0") << "true" << '\n'
+               << QStringLiteral("Version=1.0") << '\n'
                << QStringLiteral("Categories=Utility;Security;Qt;") << '\n'
                << QStringLiteral("MimeType=application/x-keepass2;") << '\n'
-               << QStringLiteral("X-GNOME-Autostart-enabled=true") << endl;
+               << QStringLiteral("X-GNOME-Autostart-enabled=true") << '\n'
+               << QStringLiteral("X-GNOME-Autostart-Delay=2") << '\n'
+               << QStringLiteral("X-KDE-autostart-after=panel") << endl;
         desktopFile.close();
     } else if (isLaunchAtStartupEnabled()) {
         QFile::remove(getAutostartDesktopFilename());